Backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib[bcrypt]
pydantic
python-multipart
python-dotenv
transformers
torch
